requirements.txt
apache-airflow
pandas
psycopg2-binary
sqlalchemy
meltano